backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
Pillow==10.2.0
python-dotenv==1.0.0
email-validator==2.1.0
opencv-python==4.9.0.80
numpy==1.26.3
tensorflow==2.17.0
